Refactored all the components related to displaying single post page in different categories.
Moved the duplicate code from SkillsPost, GiveawayPost and MonetarySupportPost components into
a new component SharedSinglePost.jsx.
