+++ frontend/src/css/App.css
@@ -19,8 +19,4 @@
 
 @import "./Components/notification/notification.css";
 
-
 @import "./Components/Post/SinglePostPage.css";
-
-@import "./Components/Post/GiveawayPost.css";
-
