--- frontend/src/css/Components/Post/GiveawayPost.css
@@ -1,160 +1,0 @@
-.single-post {
-  display: grid;
-  justify-items: center;
-}
-
-.single-post .post-pic {
-  display: grid;
-  justify-items: center;
-  box-shadow: 0 2px 20px 0 rgba(0, 0, 0, 0.6);
-  margin-bottom: 2rem;
-}
-
-.single-post .post-pic img {
-  object-fit: cover;
-  width: 100%;
-  max-height: 80vh;
-}
-
-.single-post-card {
-  width: 80vw;
-  padding: 2rem;
-  min-height: 80vh;
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  justify-items: center;
-  align-content: center;
-  box-shadow: 5px 5px 16px rgba(163, 177, 198, 0.5);
-  border-radius: 4rem;
-}
-
-.single-post-card .main {
-  width: 35rem;
-}
-
-.single-post-card .signature {
-  margin-top: 1rem;
-  display: grid;
-  grid-template-columns: auto 1fr;
-  align-items: center;
-  justify-items: start;
-}
-
-.single-post-card .signature img {
-  width: 4rem;
-  height: 4rem;
-  border-radius: 2rem;
-  margin-right: 1rem;
-}
-
-.single-post-card .signature span {
-  display: block;
-}
-
-.single-post-card .signature div {
-  height: 100%;
-  display: grid;
-  grid-row-gap: 0.3rem;
-  align-items: center;
-}
-
-.single-post-card .signature .user-name {
-  font-size: 1.5rem;
-}
-
-.single-post-card .main .post-info {
-  background-color: rgba(219, 215, 215, 0.4);
-  padding: 1rem;
-  border-radius: 2rem;
-  margin-bottom: 2rem;
-}
-
-.post-info .post-body {
-  font-size: 1.2rem;
-}
-
-.comments {
-  display: grid;
-}
-
-.no-comments {
-  color: grey;
-  font-size: 1.2rem;
-  margin: 1.5rem 0;
-}
-
-.comment-area {
-  display: grid;
-  justify-items: start;
-  align-items: start;
-}
-
-.comment-area textarea {
-  border: 1px solid rgb(190, 188, 188);
-  border-radius: 1rem;
-  width: 30rem;
-  height: 8rem;
-  padding: 0.5rem;
-  margin: 1rem 0;
-}
-
-.comments {
-  display: grid;
-  grid-row-gap: 1rem;
-}
-
-.comment-card {
-  background-color: rgba(219, 215, 215, 0.4);
-  border-radius: 2rem;
-  display: grid;
-  padding: 1rem;
-}
-
-.comment-card p {
-  font-size: 1.1rem;
-}
-
-.comments-wrapper h3 {
-  margin-bottom: 1rem;
-}
-
-.main .button-group {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.button-group .medium-button {
-  background-color: #1e6fbf;
-}
-
-.button-group button:first-child {
-  margin-right: 1rem;
-}
-
-.single-post-card .map-wrapper .map {
-  border-radius: 2rem;
-}
-
-/*
-.comment-card {
-  display: grid;
-  justify-items: center;
-  margin-left: -20rem;
-} */
-
-/* .comment-card button,
-.single-post-card button {
-  background: none;
-  height: 3rem;
-  width: 9rem;
-  margin-right: 16.5rem;
-  background-color: orange;
-  border-style: none;
-  border-radius: 1rem;
-  color: white;
-  cursor: pointer;
-} */
-
-/* .comment-card {
-  margin-left: 17rem;
-} */
