--- frontend/src/css/Components/layout/layout.css
@@ -1,34 +1,0 @@
-/* .navBar {
-  display: flex;
-  align-items: center;
-} */
-
-.navbar {
-  list-style: none;
-  color: white;
-}
-
-.navbar ul {
-  list-style: none;
-  color: white;
-  display: grid;
-  grid-template-columns: 1fr repeat(7, auto);
-  justify-items: center;
-  align-items: center;
-  grid-column-gap: 1rem;
-  padding-right: 1rem;
-}
-
-.navbar ul .logo-link {
-  justify-self: start;
-}
-
-/* .navbar ul li {
-  justify-self: flex-start;
-} */
-
-/* .navBar ul li {
-  display: flex;
-  align-items: center;
-  padding: 1rem;
-} */
